@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <!--TargetFramework>netcoreapp3.1</TargetFramework-->
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <!--AspNetCoreHostingModel>OutOfProcess</AspNetCoreHostingModel-->
     <UserSecretsId>572ea518-a8ec-422d-b9f3-8817088ece8c</UserSecretsId>
     <AssemblyVersion>1.1.0.1</AssemblyVersion>
